Update AngularJS to 1.0.6 and switch testacular out to karma.
Due to the name switch in the testacular project, start using the new project
with the name karma.  Due to this change, also upgrade AngularJS to 1.0.6
to keep e2e tests working.  Rename various files to reflect these changes,
and stop relying on karma being installed locally.  Instead rely on the
executable being available in $PATH.